improve project tracking

+++ reactjs/reactapp/copy_build.sh
@@ -3,6 +3,7 @@
 # receive distDir as argument
 
 distDir=$1
+reactBase="dist-react-base-1.0.0"
 
 key=$(($RANDOM%99999+10000))
 logFile=/var/log/project/shell_log/cmdlog.log
@@ -27,24 +28,31 @@ if [[ ! (-d "${distDir}") ]]; then
   exit;
 fi
 
+if [[ ! (-d "${reactBase}") ]]; then
+  echo "React base dir: ${reactBase}, is missing.";
+  echo "Copying file failed.";
+  exit;
+fi
+
+
 rm -rf ${distDir}/*
 
 cp build/static/js/*.js ${distDir}/
 
-mv ${distDir}/runtime-main.*.js ${distDir}/script1.js
+mv ${distDir}/runtime-main.*.js ${reactBase}/script1.js
 mv ${distDir}/main.*.chunk.js ${distDir}/script2.js
-mv ${distDir}/*.chunk.js ${distDir}/script3.js
+mv ${distDir}/*.chunk.js ${reactBase}/script3.js
 
-sed -i "2s/.*//" ${distDir}/script1.js
+sed -i "2s/.*//" ${reactBase}/script1.js
 sed -i "2s/.*//" ${distDir}/script2.js
-sed -i "3s/.*//" ${distDir}/script3.js
+sed -i "3s/.*//" ${reactBase}/script3.js
 
 
-if [[ $distDir == "dist-auth-app" ]]; then
-  cp ${distDir}/script1.js ../../../ftp-application/FTP/src/main/resources/assets/static/dist-auth-app/
-  cp ${distDir}/script2.js ../../../ftp-application/FTP/src/main/resources/assets/static/dist-auth-app/
-  cp ${distDir}/script3.js ../../../ftp-application/FTP/src/main/resources/assets/static/dist-auth-app/
-fi
+# if [[ $distDir == "dist-auth-app" ]]; then
+  # cp ${distDir}/script1.js ../../../ftp-application/FTP/src/main/resources/assets/static/dist-auth-app/
+  # cp ${distDir}/script2.js ../../../ftp-application/FTP/src/main/resources/assets/static/dist-auth-app/
+  # cp ${distDir}/script3.js ../../../ftp-application/FTP/src/main/resources/assets/static/dist-auth-app/
+# fi
 
 
 # sed -i  ''  '2s/.*//' ${distDir}/script1.js
